refactor(mail): share one row skeleton between mail and stack rows
StackListItem re-implemented MailListItem's layout rather than sharing it:
four identical class strings, the checkbox/avatar column, the full-width
switch and the date placement, all copied. They drifted, as copies do — the
stack row grew a column of its own for its chevron, which pushed its sender
and date out of line with every row around it.

Extract the skeleton both rows need into MailRow: the root (a link for a
thread, a plain div for a stack), the selection column, the sender line with
its date, subject, preview, and the trailing slot the star or chevron sits in.
Each caller fills the slots and owns nothing about the layout. MailListItem
keeps what is genuinely its own — the thread link, search highlighting,
attachments, mailbox chips, and the id MailboxView scrolls rows into view by.

The two action bars collapse into MailRowActions the same way. Its conditions
read "any member" — offer Archive while any thread is outside it, Delete only
once every thread is in Trash — which for a single thread reduces to exactly
what the mail row did before, so one component serves both. Labels name the
count only when there is more than one.

750 lines across four components becomes 658 across four, and the layout now
has one definition instead of two.
