--- frontend/src/apps/mail/components/MailListItemActions.vue
@@ -1,86 +1,0 @@
-<template>
-	<template v-if="isHovered && !isMobile">
-		<Tooltip v-for="action in actions" :key="action.label" :text="action.label">
-			<button class="action-btn" @click.stop.prevent="action.onClick">
-				<component :is="action.icon" class="icon text-ink-gray-5" />
-			</button>
-		</Tooltip>
-	</template>
-	<Tooltip :text="mail.flagged ? __('Unstar Mail') : __('Star Mail')">
-		<button class="action-btn" @click.stop.prevent="emit('setFlagged', !mail.flagged)">
-			<Star
-				class="icon text-ink-gray-5"
-				:style="mail.flagged ? 'fill: var(--ink-amber-6); color: var(--ink-amber-6)' : ''"
-			/>
-		</button>
-	</Tooltip>
-</template>
-
-<script setup lang="ts">
-import { computed } from 'vue'
-import { Archive, Mail, MailOpen, Star, Trash2 } from 'lucide-vue-next'
-import { Tooltip } from 'frappe-ui'
-
-import { useScreenSize } from '@/apps/mail/utils/composables'
-import { userStore } from '@/apps/mail/stores/user'
-
-import type { Thread } from '@/apps/mail/types'
-
-const { isHovered, mail } = defineProps<{ isHovered: boolean; mail: Thread }>()
-const emit = defineEmits(['setSeen', 'archiveThread', 'trashThread', 'deleteThread', 'setFlagged'])
-
-const { isMobile } = useScreenSize()
-const { mailboxIds } = userStore()
-
-// Prefer the row's own Archive/Trash ids when present (merged All Inboxes / cross-account search rows,
-// which can belong to a different account than the active one) so the Archive/Trash/Delete options
-// reflect where the mail actually lives. Falls back to the active account for a normal single-account view.
-const archiveId = computed(() => mail.archive ?? mailboxIds.archive)
-const trashId = computed(() => mail.trash ?? mailboxIds.trash)
-
-const actions = computed(() =>
-	[
-		{
-			label: __('Mark as Unread'),
-			onClick: () => emit('setSeen', false),
-			icon: Mail,
-			condition: mail.seen,
-		},
-		{
-			label: __('Mark as Read'),
-			onClick: () => emit('setSeen', true),
-			icon: MailOpen,
-			condition: !mail.seen,
-		},
-		{
-			label: __('Archive Thread'),
-			onClick: () => emit('archiveThread'),
-			icon: Archive,
-			condition: !mail.mailboxes.some((m) => m.mailbox_id === archiveId.value),
-		},
-		{
-			label: __('Move to Trash'),
-			onClick: () => emit('trashThread'),
-			icon: Trash2,
-			condition: !mail.mailboxes.some((m) => m.mailbox_id === trashId.value),
-		},
-		{
-			label: __('Delete Thread'),
-			onClick: () => emit('deleteThread'),
-			icon: Trash2,
-			condition: mail.mailboxes.some((m) => m.mailbox_id === trashId.value),
-		},
-	].filter((action) => action.condition),
-)
-</script>
-
-<style scoped>
-.action-btn {
-	@apply relative after:absolute after:-inset-4 after:content-[''] sm:after:-inset-1.5;
-}
-
-.action-btn:hover > * {
-	color: var(--ink-gray-8) !important;
-	stroke-width: 2 !important;
-}
-</style>
