+++ frontend/src/apps/mail/components/MailRow.vue
@@ -1,0 +1,198 @@
+<template>
+	<component
+		:is="to ? RouterLink : 'div'"
+		:to="to"
+		class="sm:hover:bg-surface-gray-1 group flex cursor-default select-none space-x-2.5 border-b px-3.5 py-2.5 sm:space-x-5 sm:px-5"
+		:class="{ '!bg-surface-blue-1': isSelected || isTouching, '!py-2': isFullWidth }"
+		@mouseenter="isHovered = true"
+		@mouseleave="isHovered = false"
+		@touchstart="onTouchStart"
+		@touchend="clearTouchTimer"
+		@touchcancel="clearTouchTimer"
+	>
+		<!-- Selection column: checkbox on desktop, sender avatar on mobile or where the list has no
+		     bulk selection. Long-pressing the row anywhere selects it on touch. -->
+		<div
+			class="flex shrink-0 items-center justify-center max-sm:w-10"
+			:class="isFullWidth ? 'h-8' : 'h-10 sm:-mt-1.5'"
+		>
+			<div
+				v-if="!isMobile && selectable"
+				class="checkbox-hitbox -m-3 cursor-pointer p-3"
+				@click.stop.prevent="emit('setSelected', !isSelected)"
+			>
+				<Checkbox :model-value="isSelected" size="md" class="pointer-events-none" />
+			</div>
+			<div
+				v-else-if="isSelected"
+				class="bg-surface-gray-10 hitbox flex h-8 w-8 shrink-0 rounded-full"
+				@click.stop.prevent="emit('setSelected', false)"
+			>
+				<Check class="text-ink-base m-auto h-5 w-5 stroke-[3px]" />
+			</div>
+			<Avatar
+				v-show="!isSelected && (isMobile || !selectable)"
+				:label="avatarLabel"
+				:image="avatarImage"
+				size="xl"
+				class="hitbox"
+				@click.stop.prevent="emit('setSelected', true)"
+			/>
+		</div>
+
+		<div class="grow truncate" :class="isFullWidth ? 'flex items-center space-x-3' : 'space-y-1'">
+			<div class="flex items-center" :class="isFullWidth ? 'w-48 shrink-0' : 'justify-between'">
+				<div class="mr-2 mt-0.5 flex items-center space-x-1.5 truncate">
+					<span v-if="unread" class="min-h-2 min-w-2 rounded-full bg-blue-500" />
+					<h3
+						class="truncate text-[15px] !font-medium sm:text-base"
+						:class="{ '!font-semibold': unread }"
+					>
+						<slot name="sender" />
+					</h3>
+					<slot name="badges" />
+				</div>
+				<MailDate v-if="!isFullWidth" :datetime :in-list="true" />
+			</div>
+
+			<h4
+				class="truncate text-sm !leading-[1.5]"
+				:class="{ italic: subjectItalic, '!text-base': isFullWidth, '!font-semibold': unread }"
+			>
+				<slot name="subject" />
+			</h4>
+
+			<div
+				class="flex items-center justify-between truncate"
+				:class="{ 'min-w-0 flex-1 !text-base': isFullWidth }"
+			>
+				<h5
+					class="text-ink-gray-5 truncate text-sm !leading-[1.5]"
+					:class="{ italic: previewItalic, '!text-base': isFullWidth }"
+				>
+					<slot name="preview" />
+				</h5>
+
+				<div v-if="!isFullWidth" class="ml-3.5 flex space-x-3.5">
+					<slot name="trailing" :is-hovered="isHovered" />
+				</div>
+			</div>
+
+			<slot name="extra" :is-full-width="isFullWidth" />
+		</div>
+
+		<div v-if="isFullWidth" class="flex w-32 shrink-0 items-center justify-end space-x-4">
+			<MailDate v-if="!isHovered" :datetime :in-list="true" />
+			<slot name="trailing" :is-hovered="isHovered" />
+		</div>
+
+		<slot />
+	</component>
+</template>
+
+<script setup lang="ts">
+import { computed, inject, ref } from 'vue'
+import { RouterLink, type RouteLocationRaw } from 'vue-router'
+import { Check } from 'lucide-vue-next'
+import { Avatar, Checkbox } from 'frappe-ui'
+
+import { useScreenSize } from '@/apps/mail/utils/composables'
+import MailDate from '@/apps/mail/components/MailDate.vue'
+
+import type { UserResource } from '@/apps/mail/types'
+
+// The shared skeleton of a row in the mail list: selection column, sender line with its date, subject,
+// preview, and the trailing slot the star or chevron sits in. It owns the layout and nothing else —
+// what goes in each slot is the caller's business.
+//
+// It exists because MailListItem and StackListItem must line up pixel-for-pixel while being different
+// things (one is a link to a thread, the other a collapsed run of them). Holding the class strings in
+// one place is what keeps them from drifting apart, which is exactly what happened when they didn't.
+const {
+	to,
+	isSelected,
+	selectable = true,
+	unread = false,
+	avatarLabel,
+	avatarImage,
+	datetime,
+	subjectItalic = false,
+	previewItalic = false,
+} = defineProps<{
+	// Renders the row as a link when set, and as a plain div when not — a thread opens, a stack expands.
+	to?: RouteLocationRaw
+	isSelected: boolean
+	// When false the desktop checkbox gives way to the sender avatar (lists without bulk selection).
+	selectable?: boolean
+	// Drives the blue dot and the bolding of the sender and subject.
+	unread?: boolean
+	avatarLabel: string
+	avatarImage?: string
+	datetime: string
+	subjectItalic?: boolean
+	previewItalic?: boolean
+}>()
+
+const emit = defineEmits<{ setSelected: [selected: boolean] }>()
+
+const user = inject('$user') as UserResource
+const { isMobile } = useScreenSize()
+
+const isHovered = ref(false)
+
+// With the reading pane hidden the list has the window to itself, so a row lays out as one wide line
+// instead of a stacked block.
+const isFullWidth = computed(() => !(user.data.show_reading_pane || isMobile.value))
+
+// Touch selection: a long press anywhere on the row selects it, as long as the finger stays put (a
+// drag is a scroll, not a press).
+let touchStartX = 0
+let touchStartY = 0
+let touchMoved = false
+let touchTimer: ReturnType<typeof setTimeout> | null = null
+
+const isTouching = ref(false)
+
+const onTouchStart = (e: TouchEvent) => {
+	touchMoved = false
+	touchStartX = e.touches[0].clientX
+	touchStartY = e.touches[0].clientY
+	isTouching.value = true
+	document.addEventListener('touchmove', onTouchMove, { passive: true })
+
+	touchTimer = setTimeout(() => {
+		if (!touchMoved) emit('setSelected', !isSelected)
+	}, 450)
+}
+
+const clearTouchTimer = () => {
+	isTouching.value = false
+	document.removeEventListener('touchmove', onTouchMove)
+
+	if (touchTimer) {
+		clearTimeout(touchTimer)
+		touchTimer = null
+	}
+}
+
+const onTouchMove = (e: TouchEvent) => {
+	const touch = e.touches[0]
+	const dx = Math.abs(touch.clientX - touchStartX)
+	const dy = Math.abs(touch.clientY - touchStartY)
+	if (dx > 10 || dy > 10) {
+		touchMoved = true
+		clearTouchTimer()
+	}
+}
+</script>
+
+<style scoped>
+.hitbox {
+	@apply relative after:absolute after:-inset-2 after:content-[''];
+}
+
+.checkbox-hitbox:hover :deep(input[type='checkbox']) {
+	@apply shadow-sm;
+	border-color: var(--outline-gray-7);
+}
+</style>
